import { terminals } from "./terminals";
