+++ lib/syntax/lex-pipe.js
@@ -6,8 +6,10 @@ var __extends = (this && this.__extends) || function (d, b) {
 };
 var _ = require("lodash");
 var frames_1 = require("../frames");
+var terminals_1 = require("./terminals");
 var tokens_1 = require("./tokens");
 var meta = _.clone(tokens_1.tokens);
+_.merge(meta, terminals_1.terminals);
 var LexPipe = (function (_super) {
     __extends(LexPipe, _super);
     function LexPipe(out) {
