+++ lib/syntax/terminals.js
@@ -1,0 +1,22 @@
+"use strict";
+var __extends = (this && this.__extends) || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+};
+var frames_1 = require("../frames");
+var lex_1 = require("./lex");
+var LexSpace = (function (_super) {
+    __extends(LexSpace, _super);
+    function LexSpace() {
+        return _super !== null && _super.apply(this, arguments) || this;
+    }
+    LexSpace.prototype.isEnd = function (char) { return char !== " "; };
+    LexSpace.prototype.makeFrame = function () { return frames_1.FrameSymbol.for(""); };
+    return LexSpace;
+}(lex_1.Lex));
+exports.LexSpace = LexSpace;
+;
+exports.terminals = {
+    " ": new LexSpace(),
+};
