+++ lib/syntax/tokens.d.ts
@@ -1,4 +1,4 @@
-import { Context, FrameSymbol } from "../frames";
+import { Context } from "../frames";
 import { Lex } from "./lex";
 import { ParseToken } from "./parse";
 export declare class LexString extends Lex {
@@ -8,9 +8,5 @@ export declare class LexString extends Lex {
 export declare class LexComment extends Lex {
     protected isEnd(char: string): boolean;
     protected makeFrame(): ParseToken;
-}
-export declare class LexSpace extends Lex {
-    protected isEnd(char: string): boolean;
-    protected makeFrame(): FrameSymbol;
 }
 export declare const tokens: Context;
