+++ lib/syntax/tokens.js
@@ -37,19 +37,7 @@ var LexComment = (function (_super) {
 }(lex_1.Lex));
 exports.LexComment = LexComment;
 ;
-var LexSpace = (function (_super) {
-    __extends(LexSpace, _super);
-    function LexSpace() {
-        return _super !== null && _super.apply(this, arguments) || this;
-    }
-    LexSpace.prototype.isEnd = function (char) { return char !== " "; };
-    LexSpace.prototype.makeFrame = function () { return frames_1.FrameSymbol.for(""); };
-    return LexSpace;
-}(lex_1.Lex));
-exports.LexSpace = LexSpace;
-;
 exports.tokens = {
-    " ": new LexSpace(),
     "#": new LexComment(),
     "“": new LexString(),
 };
