+++ src/syntax/lex-pipe.ts
@@ -1,8 +1,10 @@
 import * as _ from "lodash";
 import { Context, Frame, FrameArray, FrameLazy, FrameString, FrameSymbol, Void } from "../frames";
+import { terminals } from "./terminals";
 import { tokens } from "./tokens";
 
 const meta = _.clone(tokens);
+_.merge(meta, terminals);
 
 export class LexPipe extends Frame {
   constructor(out: Frame) {
