+++ src/syntax/terminals.ts
@@ -1,0 +1,13 @@
+import { Context, Frame, FrameString, FrameSymbol } from "../frames";
+import { Lex } from "./lex";
+import { ParseTerminal } from "./parse";
+
+export class LexSpace extends Lex {
+  protected isEnd(char: string) { return char !== " "; }
+
+  protected makeFrame() { return FrameSymbol.for(""); }
+};
+
+export const terminals: Context = {
+  " ": new LexSpace(),
+};
