+++ src/syntax/tokens.ts
@@ -22,14 +22,7 @@ export class LexComment extends Lex {
   }
 };
 
-export class LexSpace extends Lex {
-  protected isEnd(char: string) { return char !== " "; }
-
-  protected makeFrame() { return FrameSymbol.for(""); }
-};
-
 export const tokens: Context = {
-  " ": new LexSpace(),
   "#": new LexComment(),
   "“": new LexString(),
 };
